+++ include/ambition/Ambition_user_interface.hh
@@ -147,6 +147,10 @@ constexpr auto RALLY_POINT_BUTTON = Rectangle::fromPoint(
 );
 
 
+Size bitmapSize(
+  const char* bitmap
+);
+
 void drawPanel(
   const Rectangle rectangle
 );