+++ CHANGELOG.md
@@ -12,6 +12,8 @@ speaking, considered to be the player's experience.
 
 ### Changed
 
+- Show Spy combat (in Forts) or primary skill (in other Firms) level in the
+  worker list by moving the spy icon to the portrait.
 - Added thousands seperators to treasury and food store amounts, to make reading
   them easier.  (Currently, this uses groups of three with a comma as the
   seperator for everyone.  Proper localisation will be added later.)